\name{adagio-package}
\alias{adagio-package}
\alias{adagio}
\docType{package}
\title{
  Discrete and Global Optimization
}
\description{
  The R package `adagio' will provide methods and algorithms for
  discrete as well as (derivative-free) global optimization.
}
\details{
\tabular{ll}{
Package: \tab adagio\cr
Type: \tab Package\cr
Version: \tab 0.6.4\cr
Date: \tab 2015-12-07\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Hans W Borchers

  Maintainer: HwB  <hwborchers@googlemail.com>
}
\references{
  Arnold Neumaier (2010). Global Optimization.
  \url{http://www.mat.univie.ac.at/~neum/glopt.html}
}
\keyword{ package }
\seealso{
  The R-Forge project ``Optimization and solving packages''.
  \url{https://r-forge.r-project.org/projects/optimizer/}
}
\examples{
# See the functions in this package.
}
