\name{adagio-package}
\alias{adagio-package}
\alias{adagio}
\docType{package}
\title{
\packageTitle{adagio}
}
\description{
\packageDescription{adagio}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{adagio}
\packageIndices{adagio}
}
\author{
\packageAuthor{adagio}

Maintainer: \packageMaintainer{adagio}
}
